OCT-111: moved GetProductUuidsQuery
